requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/base.in
#
asgiref==3.8.1
    # via django
boto3==1.34.144
    # via -r requirements/base.in
botocore==1.34.144
    # via
    #   boto3
    #   s3transfer
django==4.2
    # via
    #   -r requirements/base.in
    #   django-storages
    #   djangorestframework
django-environ==0.10.0
    # via -r requirements/base.in
django-storages==1.14.4
    # via -r requirements/base.in
djangorestframework==3.14.0
    # via -r requirements/base.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
python-dateutil==2.9.0.post0
    # via botocore
pytz==2024.1
    # via djangorestframework
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via asgiref
urllib3==1.26.19
    # via botocore
